app: HTTPS MITM mode (engine AAR, cert install), v0.5.0

+++ mitm/tun.go
@@ -8,13 +8,13 @@ import (
 )
 
 // StartTunnel заворачивает трафик Android TUN-интерфейса (fd, полученный
-// от VpnService.Builder.establish()) в локальный прокси движка.
+// от VpnService.Builder.establish().detachFd()) в локальный прокси движка.
 // Вызывать ПОСЛЕ StartProxy — стек шлёт трафик на 127.0.0.1:8080.
-func StartTunnel(fd int, mtu int) error {
+func StartTunnel(fd int64, mtu int64) error {
 	engine.Insert(&engine.Key{
 		Device:   fmt.Sprintf("%s://%d", fdbased.Driver, fd),
 		Proxy:    "http://" + proxyAddr,
-		MTU:      mtu,
+		MTU:      int(mtu),
 		LogLevel: "error",
 	})
 	engine.Start()
